requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# AI/ML (optional)
# torch>=2.0.0
# transformers>=4.30.0

# API Clients
requests>=2.31.0
httpx>=0.25.0

# Data Processing
pandas>=2.0.0
jsonschema>=4.19.0

# CLI Tools
typer>=0.9.0
tqdm>=4.66.0
